aider: The booking date is now correctly set in the dataframe by converting it to a string before indexing.
# Aider chat conversation:

USER: when booking a date it doesnt set the true correctly in the dataframe, it just adds a new row with "true"

ASSISTANT: I see the issue. The problem is that the 'Date' column in the `booking_status` DataFrame is of type string, while the `booking_date` variable is a datetime.date object. When you try to use the `booking_date` to index into the DataFrame, it doesn't match any of the existing dates because they are different types.

To fix this, we need to convert the `booking_date` to a string in the same format as the 'Date' column before using it to index into the DataFrame.

Here is the *edit block* for the change:

```python
main.py
<<<<<<< HEAD
# Update the DataFrame with the booking information when the form is submitted
if submit_button:
    new_booking = {'Band Name': band_name, 'Booking Date': booking_date, 'Booking Time': booking_time}
    booking_data = booking_data.append(new_booking, ignore_index=True)
    booking_status.loc[booking_date, booking_time] = True
=======
# Update the DataFrame with the booking information when the form is submitted
if submit_button:
    new_booking = {'Band Name': band_name, 'Booking Date': booking_date, 'Booking Time': booking_time}
    booking_data = booking_data.append(new_booking, ignore_index=True)
    booking_date_str = booking_date.strftime('%d.%m.%Y')
    booking_status.loc[booking_status['Date'] == booking_date_str, booking_time] = True
>>>>>>> updated
```

This change will convert the `booking_date` to a string in the 'dd.mm.yyyy' format before using it to index into the `booking_status` DataFrame.